% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumal_freq.r
\name{ctLine}
\alias{ctLine}
\title{ctLine takes the index of a bin and returns a line function mapping}
\usage{
ctLine(idx, buckets)
}
\arguments{
\item{idx}{interger of bin}

\item{buckets}{histogram data from cesnsus tract as vector}
}
\value{
closure function to calculate number of homes at percentile
}
\description{
a Value to the cumalative number of tracts below that value
}
