% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumal_freq.r
\name{cumul_buckets}
\alias{cumul_buckets}
\title{cumul_buckets generate cumalative frequencies}
\usage{
cumul_buckets(x)
}
\arguments{
\item{x}{input histogram vector}
}
\value{
cumulative frequencies vector
}
\description{
converts a histogram data set to cumalative frequencies
}
\examples{
freq = [1, 4, 5, 3, 1]
cumul_buckets(freq)

}
