% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumal_freq.r
\name{findctbin}
\alias{findctbin}
\title{findctbin takes a value and find the bin tat value would exist based on histo}
\usage{
findctbin(val)
}
\arguments{
\item{val}{double value bound of percentile}
}
\value{
index of bin on census tract histogram
}
\description{
findctbin takes a value and find the bin tat value would exist based on histo
}
