% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumal_freq.r
\name{histo_findbin}
\alias{histo_findbin}
\title{histo_findbin takes a histogram and return the percentiles values}
\usage{
histo_findbin(buckets, percentiles = c(0.2, 0.8))
}
\arguments{
\item{buckets}{histogram buckets}

\item{percentiles}{vector defualts to 20%(0.2) and 80%(0.8}
}
\value{
percentile values as double
}
\description{
takes a histogram and returns the percentile values. aggregrates findbin and line together
}
