% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumal_freq.r
\name{n_line}
\alias{n_line}
\title{line returns a function that is the line equation}
\usage{
n_line(idx, buckets)
}
\arguments{
\item{idx}{index of bin integer}

\item{buckets}{histogram vector}
}
\value{
function of line to calculate
}
\description{
returns a line equation in relation to ns of histogram. This will provide a closure that will return the value of the houses.
}
