% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumal_freq.r
\name{perc_findbin}
\alias{perc_findbin}
\title{findbin takes a percentile and returnsa value of % n}
\usage{
perc_findbin(perc, buckets)
}
\arguments{
\item{perc}{percentiles vector}

\item{buckets}{histogram vector}
}
\value{
integer index of the bin on histo
}
\description{
the target histogramic data buckets and returns which buckets would perc exist in that histogram
}
\examples{
histogram_freq = [10, 20, 40, 10, 0, 0, 1]

percentile_20th_idx = perc_findbin(0.5, histogram_freq)

median_idx = perc_bin(0.5, histogram_freq)
median_idx == median(histogram_freq)
}
