% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumal_freq.r
\name{percs_MSA_2000}
\alias{percs_MSA_2000}
\title{perc_MSA_2000 is a helper function to compenstate for census 2000}
\usage{
percs_MSA_2000(ct_data, percentiles = c(0.2, 0.8))
}
\arguments{
\item{ct_data}{data.frame of census tract data}

\item{percentiles}{vector of percentiles}
}
\value{
percentiles values vector
}
\description{
Census 2000 did not include MSA home value data but can generate those values from census tract data
}
