% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumal_freq.r
\name{return_bucket}
\alias{return_bucket}
\title{return_bucket return the text label of the bucket}
\usage{
return_bucket(idx)
}
\arguments{
\item{idx}{the index of the bucket to be returned}
}
\value{
the text label of the bucket at index idx
}
\description{
this is used for testing to make sure that value falls inside
}
